requirements.txt
asgiref==3.7.2
Django==4.2.2
psycopg==3.1.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
django-debug-toolbar==1.11
djangorestframework==3.14.0
